s10

+++ components/axaddrspace/src/address_space/mod.rs
@@ -96,7 +96,7 @@ impl<H: PagingHandler> AddrSpace<H> {
             return ax_err!(InvalidInput, "address not aligned");
         }
 
-        let offset = start_vaddr.as_usize() - start_paddr.as_usize();
+        let offset = (start_vaddr.as_usize() as isize) - (start_paddr.as_usize() as isize);
         let area = MemoryArea::new(start_vaddr, size, flags, Backend::new_linear(offset));
         self.areas
             .map(area, &mut self.pt, false)
